@@ -32,7 +32,7 @@
     "@types/react": "^16.9.34",
     "@types/react-dom": "^16.9.6",
     "@types/styled-components": "^5.1.2",
-    "eslint": "^6.8.0",
+    "eslint": "^7.7.0",
     "eslint-plugin-react": "^7.19.0",
     "npm-watch": "^0.7.0",
     "puppeteer": "^5.2.1",
@@ -41,9 +41,11 @@
     "styled-components": "^5.1.0",
     "ts-loader": "^6.2.2",
     "ts-node": "^9.0.0",
-    "typescript": "^3.8.3"
+    "typescript": "^4.0.2"
   },
   "devDependencies": {
-    "@types/node": "^14.6.2"
+    "@types/node": "^14.6.2",
+    "@typescript-eslint/eslint-plugin": "^4.0.1",
+    "@typescript-eslint/parser": "^4.0.1"
   }
 }
